requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
numpy>=1.23.0
hydra-core>=1.3.0
omegaconf>=2.3.0
matplotlib>=3.7.0
Pillow>=9.5.0
tensorboard>=2.12.0
